Brings in the interpreter->SIC tier-up machinery (interpreter PICs as type
feedback, nmethod routing, DI-aware PIC constraints, loop back-edge counting,
return-trap disambiguation) plus the ui2 zoom/jump-scroll features, the morph
roles system, and the expanderMorph/windowMorph module factoring.

Conflict resolutions of note:
- interpreter.{cpp,hh}: kept our park_at_kill_target kill/retry (6/2026 work,
  supersedes the older cherry-picked kill handling, which is deleted) and our
  NLR-through-C interpret() result fix; took the chained lookup_in_progress
  registration (push/pop) and all SIC hunks; find_interpreter_for_frame now
  composes our walk-hint mechanism with their contains()/stackFor() fast-outs.
- process.cpp: took their interpreter-only guard around the killer-state
  reset and their traceV-gated bottom-frame diagnostic; KEPT our
  is_done_with_killing_or_deoptimizing dead-killVF semantics (returns
  "not the target", theirs returns "done") -- flagged for reconciliation.
- spaceSize.hh: kept our 512GB-macOS / 24GB-visionOS heap layout (newer
  measured knowledge than their 32/48GB choice).
- frame.{hh,cpp}, conversion.cpp: took their widened 64-bit guards (the
  sentinel distinction and recursive-C++ NLR shape hold in SIC builds too).
- objects/: kept our classic-world lineage (categories, license lines,
  windowCanvasPrototypeForDisplay: XQuartz fallback); took their zoom +
  roles + event-drain features; allUI2 gains only the two factored modules
  that exist here (expanderMorph, windowMorph -- not the four GitHub-world
  modules our tree doesn't carry); viewScrollMorph stays deleted.

+++ vm64/src/any/prims/glueDefs.hh
@@ -794,7 +794,7 @@ typedef class {
 
 # define proxy_null_or_errno_ret(dummy, type_seal)                            \
     proxyOop result = proxyOop(s_arg(res));                                   \
-    result->set_pointer((void *)res);                                         \
+    result->set_pointer((void *)(unsigned long)res);                                         \
     result->set_type_seal((void *)type_seal);                                 \
     RETURN(result)
